Reorganiza o código movendo `create_id_file` para `isis_cmd`
