+++ dsm/extdeps/isis_migration/isis_cmds.py
@@ -39,3 +39,67 @@ def _get_document_isis_db(pid):
     while "finished" not in read_file(finished_file_path):
         pass
     return output_file_path
+
+
+def create_id_file(db_file_path, id_file_path=None):
+    """
+    Generates ID file `id_file_path` of a ISIS database `db_file_path`
+
+    Parameters
+    ----------
+    db_file_path: str
+        path of an ISIS database without extension
+    id_file_path: str
+        path of the ID file to be created
+
+    Returns
+    -------
+    str
+        id_file_path
+
+    Raises
+    ------
+    exceptions.MissingCisisPathEnvVarError
+    exceptions.CisisPathNotFoundMigrationError
+    exceptions.MissingI2IdCommandPathEnvVarError
+    exceptions.IsisDBNotFoundError
+    PermissionError
+    FileNotFoundError
+    """
+    # obtém CISIS_PATH
+    cisis_path = configuration.get_cisis_path()
+
+    # check if the utilitary i2id exists
+    i2id_cmd = os.path.join(cisis_path, "i2id")
+    if not os.path.isfile(i2id_cmd):
+        raise exceptions.MissingI2IdCommandPathEnvVarError(
+            f"Not found: {i2id_cmd}"
+        )
+
+    # check if the isis database exists
+    if not os.path.isfile(db_file_path + ".mst"):
+        raise exceptions.IsisDBNotFoundError(f"Not found {db_file_path}.mst")
+
+    if id_file_path is None:
+        # create id_file in a temp folder
+        id_file_path = create_temp_file(os.path.basename(db_file_path))
+    else:
+        # create the destination folder
+        dirname = os.path.dirname(id_file_path)
+        if not os.path.isdir(dirname):
+            os.makedirs(dirname)
+
+        # delete the id_file_path
+        write_file(id_file_path, "")
+
+    # execute i2id db > id_file_path
+    os.system(f"{i2id_cmd} {db_file_path} > {id_file_path}")
+
+    # check if id_file_path is valid
+    try:
+        content = read_file(id_file_path, encoding="iso-8859-1")
+    except FileNotFoundError:
+        return None
+    else:
+        return id_file_path
+
